+++ client/tws_foundation_client/lib/src/services/bases/trailers_externals_service_base.dart
@@ -12,4 +12,7 @@ abstract class TrailersExternalsServiceBase extends CSMServiceBase {
   /// Transaction to generate a set view object.
   Effect<SetViewOut<TrailerExternal>> view(SetViewOptions<TrailerExternal> options, String auth);
 
+  Effect<SetBatchOut<TrailerExternal>> create(List<TrailerExternal> trailers, String auth);
+
+  Effect<RecordUpdateOut<TrailerExternal>> update(TrailerExternal trailer, String auth);
 }