+++ client/tws_foundation_client/test/definition/trailer_service_test.dart
@@ -6,19 +6,30 @@ import 'package:tws_foundation_client/tws_foundation_client.dart';
 
 void main() {
   late TrailersServiceBase service;
+
   late SetViewOut<Trailer> viewMock;
+  late SetBatchOut<Trailer> createMock;
+  late RecordUpdateOut<Trailer> updateMock;
+  
   late SetViewOptions<Trailer> options;
+  late List<Trailer> trailers;
 
   setUp(
     () {
       List<SetViewOrderOptions> noOrderigns = <SetViewOrderOptions>[];
       options = SetViewOptions<Trailer>(false, 10, 1, null, noOrderigns, <SetViewFilterNodeInterface<Trailer>>[]);
       viewMock = SetViewOut<Trailer>(<Trailer>[], 1, DateTime.now(), 3, 0, 20);
-
+      createMock = SetBatchOut<Trailer>(<Trailer>[], <SetOperationFailure<Trailer>>[], 0, 0, 0, false);
+      updateMock = RecordUpdateOut<Trailer>(Trailer.a(), Trailer.a());
+      trailers = <Trailer>[
+        Trailer.a(),
+      ];
       Client mockClient = MockClient(
         (Request request) async {
           JObject jObject = switch (request.url.pathSegments.last) {
             'view' => SuccessFrame<SetViewOut<Trailer>>('qTracer', viewMock).encode(),
+            'create' => SuccessFrame<SetBatchOut<Trailer>>('qTracer', createMock).encode(),
+            'update' => SuccessFrame<RecordUpdateOut<Trailer>>('qTracer', updateMock).encode(),
             _ => <String, dynamic>{},
           };
 
@@ -62,5 +73,49 @@ void main() {
       expect(passed, true, reason: 'expected the service returned a success');
     },
     timeout: Timeout.factor(5),
+  );
+
+  test(
+    'Create',
+    () async {
+      MainResolver<SetBatchOut<Trailer>> fact = await service.create(trailers, '');
+      bool pased = false;
+      fact.resolve(
+        decoder: (JObject json) => SetBatchOut<Trailer>.des(json, Trailer.des),
+        onConnectionFailure: () {},
+        onFailure: (FailureFrame failure, int status) {
+          throw failure;
+        },
+        onSuccess: (SuccessFrame<SetBatchOut<Trailer>> success) {
+          pased = true;
+        },
+        onException: (Object exception, StackTrace trace) {
+          throw exception;
+        },
+      );
+      expect(pased, true);
+    },
+  );
+
+  test(
+    'Update',
+    () async {
+      MainResolver<RecordUpdateOut<Trailer>> fact = await service.update(Trailer.a(), '');
+      bool pased = false;
+      fact.resolve(
+        decoder: (JObject json) => RecordUpdateOut<Trailer>.des(json, Trailer.des),
+        onConnectionFailure: () {},
+        onFailure: (FailureFrame failure, int status) {
+          throw failure;
+        },
+        onSuccess: (SuccessFrame<RecordUpdateOut<Trailer>> success) {
+          pased = true;
+        },
+        onException: (Object exception, StackTrace trace) {
+          throw exception;
+        },
+      );
+      expect(pased, true);
+    },
   );
 }