+++ client/tws_foundation_client/test/integration/yard_logs_service_test.dart
@@ -95,12 +95,15 @@ void main() {
       1, //status
       0, //common
       1, //carrier
+      null,
       null, //model
-      1, //maintenance
+      null, //maintenance
+      null,
       null,
       trailerCommon, //trailerCommonNavigation
       null, //vehiculeModelNavigation
       null, //statusNavigation
+      null,
       <Plate>[plateMX2, plateUSA2] //plates
     );
     TruckCommon truckCommon = TruckCommon(